requirements.txt
streamlit==1.35.0
pdfminer.six==20221105
pytesseract==0.3.10          # bare hvis du vil ha OCR
pillow                       # trengs av pytesseract
ollama==0.1.8
python-dotenv==1.0.1
weasyprint==62.2             # valgfritt, pdf-eksport
pymupdf==1.24.3
stqdm